cloudflare-dyndns: Restructure directory to golang best practices
Fix linking issue for example config in README.md.
Fix missing .golangci trigger for workflow.

Directory layout is according to:
https://github.com/golang-standards/project-layout
Add description of the layout to main README.md
