+++ .github/workflows/build-cloudflare-dyndns.yaml
@@ -8,7 +8,6 @@ on:
     branches: [main]
     paths:
       - "apps/cloudflare-dyndns/Dockerfile"
-      - "apps/cloudflare-dyndns/*.go"
       - "apps/cloudflare-dyndns/go.mod"
       - "apps/cloudflare-dyndns/go.sum"
       - "apps/cloudflare-dyndns/**/*.go"
@@ -16,11 +15,11 @@ on:
     branches: [main]
     paths:
       - "apps/cloudflare-dyndns/Dockerfile"
-      - "apps/cloudflare-dyndns/*.go"
       - "apps/cloudflare-dyndns/go.mod"
       - "apps/cloudflare-dyndns/go.sum"
       - "apps/cloudflare-dyndns/**/*.go"
       - ".github/workflows/build-cloudflare-dyndns.yaml"
+      - ".golangci.yml"
 
 jobs:
   golangci:
