+++ apps/cloudflare-dyndns/pkg/client/client_test.go
