+++ apps/cloudflare-dyndns/pkg/client/errors.go
