+++ apps/cloudflare-dyndns/pkg/client/utils.go
