+++ apps/cloudflare-dyndns/pkg/config/testdata/invalid-config-1.yaml
