+++ apps/cloudflare-dyndns/pkg/config/testdata/not-a-config.txt
