+++ apps/cloudflare-dyndns/pkg/config/testdata/valid-config-1.yaml
