+++ apps/cloudflare-dyndns/pkg/server/server.go
@@ -9,8 +9,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/heathcliff26/containers/apps/cloudflare-dyndns/client"
-	"github.com/heathcliff26/containers/apps/cloudflare-dyndns/config"
+	"github.com/heathcliff26/containers/apps/cloudflare-dyndns/pkg/client"
+	"github.com/heathcliff26/containers/apps/cloudflare-dyndns/pkg/config"
 )
 
 type Server struct {
