+++ apps/cloudflare-dyndns/pkg/server/utils.go
