+++ apps/cloudflare-dyndns/pkg/server/utils_test.go
